requirements.txt
anyio==4.9.0
blinker==1.9.0
anyio==4.9.0
blinker==1.9.0
certifi==2025.1.31
click==8.1.8
Flask==3.1.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
python-dotenv==1.1.0
sniffio==1.3.1
typing_extensions==4.13.2
upstash-redis==1.4.0
Werkzeug==3.1.3